AOSCOS: LOONGSON3: 45drm: limit DRM module installation to amdgpu, loongson, and radeon
If we are using mips64el (assumed to be MIPS-based Loongson), only install
modules for loongson, radeon, and amdgpu as these are the only three DRM
drivers supported by these platforms. We need to minimise the sizes of the
initramfs so that it fits within the RAM allocated by the system firmware.

Signed-off-by: Mingcong Bai <jeffbai@aosc.io>